Trait sui_exex::launcher::LaunchExEx

pub trait LaunchExEx: Send {
    // Required method
    fn launch(
        self,
        ctx: ExExContext,
    ) -> impl Future<Output = Result<impl Future<Output = Result<()>> + Send>> + Send;
}

A trait for launching an ExEx.

Required Methods

fn launch( self, ctx: ExExContext, ) -> impl Future<Output = Result<impl Future<Output = Result<()>> + Send>> + Send

Launches the ExEx.

The ExEx should be able to run independently and emit events on the channels provided in the ExExContext.

Object Safety

This trait is not object safe.